connect to database

+++ db/migrate/20150401084134_create_ratings.rb
@@ -1,8 +1,8 @@
 class CreateRatings < ActiveRecord::Migration
   def change
     create_table :ratings do |t|
-      t.decimal :rating
-      t.decimal :wtfrating
+      t.integer :rating
+      t.integer :wtfrating
       t.string :identifier
 
       t.timestamp
